+++ css-inline/src/html/mod.rs
@@ -8,5 +8,5 @@ mod selectors;
 mod serializer;
 
 pub(crate) use self::selectors::Specificity;
-pub(crate) use document::Document;
+pub(crate) use document::{Document, DEFAULT_HTML_TREE_CAPACITY};
 pub(crate) use node::NodeId;
